Diet Classifier/nutrients_classifier/src/predict.py
import argparse
import numpy as np
import tensorflow as tf
from pathlib import Path
import json
from PIL import Image

from nutrition_engine import NutritionEngine

IMG_SIZE = 224

def load_classifier(model_path):
    if not Path(model_path).exists():
        raise FileNotFoundError(f"Classifier model not found at {model_path}. Please train first.")
    return tf.keras.models.load_model(model_path)

def load_portion_estimator(model_path):
    if not Path(model_path).exists():
        print(f"Warning: Portion estimator not found at {model_path}. Using fallback rule-based sizes.")
        return None
    return tf.keras.models.load_model(model_path)

def preprocess_image(image_path):
    img = Image.open(image_path).convert('RGB')
    img_resized = img.resize((IMG_SIZE, IMG_SIZE))
    img_array = np.array(img_resized) / 255.0
    return np.expand_dims(img_array, axis=0)

def main():
    parser = argparse.ArgumentParser(description="Food Nutrition Predictor")
    parser.add_argument("image_path", type=str, help="Path to the food image")
    args = parser.parse_args()

    # Paths
    base_dir = Path(__file__).resolve().parent.parent
    classifier_path = base_dir / "models" / "food_classifier_custom.h5"
    portion_path = base_dir / "models" / "portion_estimator.h5"
    label_map_path = base_dir / "data" / "processed" / "label_map.json"

    if not Path(args.image_path).exists():
        print(f"Image not found: {args.image_path}")
        return

    # Load Label Map
    if not label_map_path.exists():
        raise FileNotFoundError("label_map.json not found. Run preprocess.py first.")
    with open(label_map_path) as f:
        label_map = json.load(f)

    print("Loading models...")
    classifier = load_classifier(classifier_path)
    portion_estimator = load_portion_estimator(portion_path)
    engine = NutritionEngine(base_dir / "data")

    print(f"Processing image: {args.image_path}...")
    img_tensor = preprocess_image(args.image_path)

    # 1. Classification
    preds = classifier.predict(img_tensor, verbose=0)[0]
    
    # Get Top-3
    top_3_idx = preds.argsort()[-3:][::-1]
    
    # We will assume single plate item for CLI, or multi-item if object detection was implemented.
    # The prompt asked for "Top-3 predictions with confidence scores" and "Handles multi-food plates by summing". 
    # Since we built a full image classifier (not object detection YOLO), 
    # we treat Top-1 as the primary food unless we do multi-label.
    # We will just yield the top 1 for nutrition calculation to be logically sound, 
    # but print Top 3.
    
    primary_food_id = top_3_idx[0]
    primary_food_name = label_map[str(primary_food_id)]
    primary_conf = preds[primary_food_id] * 100

    print("\n🍽️  DETECTED FOODS:\n")
    for i, idx in enumerate(top_3_idx):
        food_name = label_map[str(idx)]
        conf = preds[idx] * 100
        
        # 2. Portion Estimation
        if portion_estimator is not None and i == 0:
            est_grams = portion_estimator.predict(img_tensor, verbose=0)[0][0]
            est_grams = max(50.0, float(est_grams)) # Minimum 50g
        else:
            # Rule based fallback
            # Usually standard serving is 200g
            est_grams = 200.0 if i == 0 else 0.0
            
        if i == 0:
            primary_grams = est_grams
            
        gram_str = f" — Est. {est_grams:.0f}g" if i == 0 else ""
        print(f"{food_name.title().replace('_', ' ')} (confidence: {conf:.1f}%){gram_str}")

    # 3. Nutrition Calculation
    detections = [{'food': primary_food_name, 'grams': primary_grams}]
    result = engine.compute_plate_nutrition(detections)
    
    # 4. Report
    print("\n📊 TOTAL NUTRITION SUMMARY:")
    print(engine.format_nutrition_label(result['total']))

if __name__ == "__main__":
    main()
