Diet Classifier/nutrients_classifier/requirements.txt
tensorflow>=2.12
gradio
opencv-python
pandas
numpy
matplotlib
scikit-learn
kaggle
pillow
seaborn
tqdm
tensorboard
fuzzywuzzy
python-Levenshtein
reportlab
